refractor: remove the callback into a hook
